Move recyclerView instantiator functions
