+++ viewdsl-recyclerview/src/main/java/splitties/viewdsl/recyclerview/SrollWrapping.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package splitties.viewdsl.recyclerview.experimental
+package splitties.viewdsl.recyclerview
 
 import android.support.annotation.IdRes
 import android.view.View
